Changes on non-SQL commands: 'filename': Bugfix, ignored the given database id (always returned the name of the first database)
'status':
When called with no arguments (or 0 as dbid) mksqlite('status') writes the statuses for each database id to the command window, as before. If called as function (status = mksqlite(...)), a cell array with respective strings "OPEN" or "CLOSED" will be returned with no output to the command window.
When called with a database id other than 0, the status only for this id is written to command window. If called as function a respective string "OPEN" or "CLOSED" will be returned, with no output to the command window.